# contracts/openapi.v0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeLess API",
    "version": "0.0.0",
    "description": "CodeLess platform API. Authentication uses a host-only session cookie and CSRF tokens. Task state is derived from real tool results; publication requires an authenticated user and a verified immutable version.",
    "license": { "name": "Proprietary", "url": "https://codeless.dev/license" }
  },
  "servers": [{ "url": "/", "description": "Same-origin platform gateway" }],
  "security": [{ "sessionCookie": [] }],
...
  "paths": {
...
    "/api/v0/applications/{applicationId}/versions/{versionId}/preview-credentials": {
      "parameters": [
        {
          "name": "applicationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "versionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "createPreviewCredential",
        "summary": "Issue a two-minute credential for a registered verified version",
        "description": "Requires host-only platform session, CSRF and ownership. Exact committed build/source/artifact must already be registered by the trusted resident gateway. Response is no-store; url is a temporary bearer capability and must not be logged or persisted.",
        "tags": [
          "preview"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped credential, no-store",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/v0/preview-credential.schema.json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/v0/error.schema.json"
                }
              }
            }
          },
          "403": {
            "description": "CSRF token rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/v0/error.schema.json"
                }
              }
            }
          },
          "404": {
            "description": "Not owned or resource mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/v0/error.schema.json"
                }
              }
            }
          },
          "409": {
            "description": "Version/build not verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/v0/error.schema.json"
                }
              }
            }
          },
          "503": {
            "description": "Gateway unavailable, unregistered version, or invalid deployment configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/v0/error.schema.json"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID",
        "description": "Host-only HttpOnly session cookie; Secure by default and SameSite=Lax. Local plain HTTP may set CODELESS_COOKIE_SECURE=false on the API."
      }
    },
    "parameters": {
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": true, "description": "Token returned by GET /api/v0/auth/csrf for this session.", "schema": { "type": "string", "pattern": "^[0-9a-f]{64}$" } },
...
    },
...
  }
}